fix(kdropdown): rename component [KHCP-8879]
