+++ src/components/KDropdown/KDropdown.vue
@@ -90,7 +90,7 @@ const props = defineProps({
     validator: (value: ButtonAppearance) => {
       // @ts-ignore
       if (value === 'menu' || value === 'selectionMenu') {
-        console.warn('KDropdownMenu: some props have changed. Please see the migration guide for more details: https://alpha--kongponents.netlify.app/guide/migrating-to-version-9.html#kdropdownmenu')
+        console.warn('KDropdown: some props have changed. Please see the migration guide for more details: https://alpha--kongponents.netlify.app/guide/migrating-to-version-9.html#kdropdownmenu')
       }
 
       return Object.values(ButtonAppearances).indexOf(value) !== -1
@@ -145,7 +145,7 @@ const props = defineProps({
     default: '',
     validator: (value: string) => {
       if (value) {
-        console.warn('KDropdownMenu: label prop is deprecated. Please use `triggerText` prop instead. Please see the migration guide for more details: https://alpha--kongponents.netlify.app/guide/migrating-to-version-9.html#kdropdownmenu')
+        console.warn('KDropdown: label prop is deprecated. Please use `triggerText` prop instead. Please see the migration guide for more details: https://alpha--kongponents.netlify.app/guide/migrating-to-version-9.html#kdropdownmenu')
       }
 
       return true
