+++ src/components/index.ts
@@ -18,8 +18,9 @@ export { default as KCheckbox } from './KCheckbox/KCheckbox.vue'
 export { default as KCodeBlock } from './KCodeBlock/KCodeBlock.vue'
 export { default as KCollapse } from './KCollapse/KCollapse.vue'
 export { default as KDateTimePicker } from './KDateTimePicker/KDateTimePicker.vue'
-export { default as KDropdownMenu } from './KDropdownMenu/KDropdownMenu.vue'
-export { default as KDropdownItem } from './KDropdownMenu/KDropdownItem.vue'
+export { default as KDropdown } from './KDropdown/KDropdown.vue'
+export { default as KDropdownItem } from './KDropdown/KDropdownItem.vue'
+export { default as KDropdownMenu } from './KDropdown/KDropdown.vue' // KDropdown alias
 export { default as KEmptyState } from './KEmptyState/KEmptyState.vue'
 export { default as KFileUpload } from './KFileUpload/KFileUpload.vue'
 export { default as KIcon } from './KIcon/KIcon.vue'
